🔀 Merge dev → main

+++ Services/Interfaces/IDiceGameService.cs
@@ -4,5 +4,5 @@ namespace TeleCasino.DiceGameService.Services.Interface;
 
 public interface IDiceGameService
 {
-    Task<DiceResult> PlayGameAsync(int wager, string betArg, int gameSessionId);
+    Task<DiceResult> PlayGameAsync(decimal wager, string betArg, int gameSessionId);
 }
